Update project structure and enhance documentation
- Disabled compiler warning in `Book.cs`.
- Added dependency injection and configuration in `Program.cs`, including a new `Setup` method.
- Enhanced `RangeHelpers.cs` with XML documentation and adjusted `EndIndex` calculation.
- Improved `WineOperations.cs` with XML documentation and simplified context initialization.
- Updated `WineContext.cs` to use a singleton for connection strings and added data seeding logic.
- Added XML documentation in `Container.cs` and removed `MonthIndex`.
- Simplified constructor in `WineGroupItem.cs` and added XML documentation.
- Upgraded target framework to .NET 9.0 and updated package references in `WineConsoleApp.csproj`.
- Revised `readme.md` for new setup instructions and added `appsettings.json` for configuration management.
- Introduced `DataConnections.cs` for managing data connection configurations.
- Added `Populate.sql` script for database creation and initial data seeding.

by Karen

+++ WineConsoleApp/Program.cs
@@ -1,4 +1,5 @@
-﻿using WineConsoleApp.Classes;
+﻿using ConsoleConfigurationLibrary.Classes;
+using WineConsoleApp.Classes;
 using WineConsoleApp.Data;
 
 namespace WineConsoleApp;
@@ -9,8 +10,12 @@ internal partial class Program
     {
         using var context = new WineContext();
 
-        context.Database.EnsureDeleted();
-        context.Database.EnsureCreated();
+        if (EntitySettings.Instance.CreateNew)
+        {
+            context.Database.EnsureDeleted();
+            context.Database.EnsureCreated();
+        }
+
         WineOperations.Run();
 
         Line();
